Write pipeline incorrectly, with race conditions.
Build with race `go build --race && ./pipeline`, and you see a lot
of issues with race conditions. There's a better way to write this
program to avoid this problem. One way is using the producer and
consumer pattern.

The naive solution would be to use mutex. We aren't going to do
that for this workshop. This design is not elegant. It ends up
requiring a lot of global variables and can get messy.